Consolidated tocStore and fixed update bug
