+++ src/components/RightSidebar/RightSidebar.astro
@@ -1,5 +1,5 @@
 ---
-import TableOfContents from "./TableOfContents"
+import TableOfContents from "~/components/TableOfContents"
 import MoreMenu from "./MoreMenu.astro"
 import { Feedback } from "../PageContent/Feedback"
 import { MarkdownHeading } from "astro"
