+++ src/components/TableOfContents/tocStore.ts
@@ -2,5 +2,5 @@ import { atom } from "nanostores"
 
 let updateCounter = 0
 
-export const shouldUpdateToc = atom<number | undefined>()
+export const shouldUpdateToc = atom<number>(updateCounter)
 export const updateTableOfContents = () => shouldUpdateToc.set(++updateCounter)
