+++ src/layouts/DocsLayout.astro
@@ -11,7 +11,7 @@ import Footer from "~/components/Footer/Footer.astro"
 import { BaseFrontmatter } from "~/content/config"
 import WhatsNext from "~/components/PageContent/WhatsNext.astro"
 import { MarkdownHeading } from "astro"
-import MobileToc from "~/components/MobileToc/MobileToc"
+import MobileToc from "~/components/TableOfContents/MobileToc/MobileToc"
 
 interface Props {
   frontmatter: BaseFrontmatter
